[DR-90] refactor: enhance FormStep component with conditional rendering
- Added activeStep prop for conditional step rendering
- Extracted useFormStep hook to separate file for better organization
- Simplified FormStep logic by removing complex validation/submission handling
- Updated imports and exports for cleaner API
