+++ fe/cms/src/components/FormStep/FormStep.tsx
@@ -1,39 +1,44 @@
 import React, { useState } from "react";
-import { useMultiStepFormContext } from "../MultiStepForm/context";
+import { useMultiStepFormContext } from "../MultiStepForm";
 import { useFormDataContext } from "../FormDataContext";
+import { FormStepContext } from "./useFormStep";
 
-interface FormStepProps<TData = any> {
+interface FormStepProps {
   stepId: string;
   children: React.ReactNode;
-  
-  // Validation - runs before submission
-  onValidate?: () => boolean | Promise<boolean>;
-  
-  // Data processing/submission logic
-  onSubmit?: (data: TData) => Promise<void>;
-  
+
   // Custom next step logic (optional)
   onNext?: () => void;
-  
+
   // Skip this step based on conditions
   skipIf?: () => boolean;
-  
+
+  // Only render on specific step
+  activeStep?: number;
+
   // Loading states
   isLoading?: boolean;
 }
 
-export function FormStep<TData>({ 
-  stepId, 
-  children, 
-  onValidate, 
-  onSubmit,
+export function FormStep({
+  stepId,
+  children,
   onNext,
   skipIf,
-  isLoading = false
-}: FormStepProps<TData>) {
-  const { bumpStep, goBackStep } = useMultiStepFormContext((state) => state);
-  const { getData, setData } = useFormDataContext((state) => state);
+  activeStep,
+  isLoading = false,
+}: FormStepProps) {
+  const {
+    bumpStep,
+    goBackStep,
+    step: currentStep,
+  } = useMultiStepFormContext((state) => state);
   const [isProcessing, setIsProcessing] = useState(false);
+
+  // Don't render if activeStep is specified and doesn't match current step
+  if (activeStep !== undefined && currentStep !== activeStep) {
+    return null;
+  }
 
   const handleNext = async () => {
     if (skipIf?.()) {
@@ -42,33 +47,15 @@ export function FormStep<TData>({
     }
 
     setIsProcessing(true);
-    
-    try {
-      // Run validation first
-      if (onValidate && !(await onValidate())) {
-        setIsProcessing(false);
-        return;
-      }
-      
-      // Get current data and run submission logic
-      const currentData = getData<TData>(stepId);
-      if (onSubmit) {
-        await onSubmit(currentData);
-      }
-      
-      // Custom next logic or default bump
-      if (onNext) {
-        onNext();
-      } else {
-        bumpStep();
-      }
-      
-    } catch (error) {
-      console.error(`Step ${stepId} failed:`, error);
-      // Handle error (maybe show toast)
-    } finally {
-      setIsProcessing(false);
+
+    // Custom next logic or default bump
+    if (onNext) {
+      onNext();
+    } else {
+      bumpStep();
     }
+
+    setIsProcessing(false);
   };
 
   const handlePrevious = () => {
@@ -76,8 +63,8 @@ export function FormStep<TData>({
   };
 
   return (
-    <FormStepProvider 
-      stepId={stepId} 
+    <FormStepProvider
+      stepId={stepId}
       onNext={handleNext}
       onPrevious={handlePrevious}
       isProcessing={isProcessing || isLoading}
@@ -87,15 +74,6 @@ export function FormStep<TData>({
   );
 }
 
-interface FormStepContextValue {
-  stepId: string;
-  onNext: () => Promise<void>;
-  onPrevious: () => void;
-  isProcessing: boolean;
-}
-
-const FormStepContext = React.createContext<FormStepContextValue | null>(null);
-
 export const FormStepProvider: React.FC<{
   stepId: string;
   onNext: () => Promise<void>;
@@ -104,16 +82,10 @@ export const FormStepProvider: React.FC<{
   children: React.ReactNode;
 }> = ({ stepId, onNext, onPrevious, isProcessing, children }) => {
   return (
-    <FormStepContext.Provider value={{ stepId, onNext, onPrevious, isProcessing }}>
+    <FormStepContext.Provider
+      value={{ stepId, onNext, onPrevious, isProcessing }}
+    >
       {children}
     </FormStepContext.Provider>
   );
-};
-
-export const useFormStep = () => {
-  const context = React.useContext(FormStepContext);
-  if (!context) {
-    throw new Error('useFormStep must be used within FormStepProvider');
-  }
-  return context;
 };
