+++ fe/cms/src/components/FormStep/FormStepNavigation.tsx
@@ -1,10 +1,10 @@
 import { ArrowLeft, ArrowRight } from "@phosphor-icons/react";
 import { Forms, FormButton } from "@/components";
-import { useFormStep } from "./FormStep";
+import { useFormStep } from "./useFormStep";
 
 export const FormStepNavigation = () => {
   const { onNext, onPrevious, isProcessing } = useFormStep();
-  
+
   return (
     <div className="w-full">
       <Forms.HorizontalLayout>
@@ -16,12 +16,12 @@ export const FormStepNavigation = () => {
         >
           <ArrowLeft /> 上一步
         </Forms.Button>
-        <FormButton 
-          extendClass="bg-white" 
+        <FormButton
+          extendClass="bg-white"
           onClick={onNext}
           disabled={isProcessing}
         >
-          {isProcessing ? 'Processing...' : '下一步'} <ArrowRight />
+          {isProcessing ? "Processing..." : "下一步"} <ArrowRight />
         </FormButton>
       </Forms.HorizontalLayout>
     </div>
