+++ fe/cms/src/components/FormStep/index.ts
@@ -1,2 +1,3 @@
-export { FormStep, FormStepProvider, useFormStep } from './FormStep';
-export { FormStepNavigation } from './FormStepNavigation';
+export { FormStep, FormStepProvider } from "./FormStep";
+export { FormStepNavigation } from "./FormStepNavigation";
+export { useFormStep } from "./useFormStep";
